Filter users on the server if possible
JupyterHub 1.3.0 supports server side filtering of the users with
servers in a given state [1]. This should perform better when culling
a large number of users since the filtering can happen in the database
rather than in the client side script on _all_ users.

The idle-culler is currently just calling `GET /users` and filtering
out users with pending servers on the client side.

This change checks to see if the hub API is new enough for server
side `state` filtering and if so, filters the initial set of users
to only those who have any ready servers (running, not pending).
Furthermore, if `--cull-users` is used we make a second call to
`GET /users?state=inactive` to get all users who have _no_ active
servers since they would have been filtered out of the initial
set (`GET /users?state=ready`).

If the jupyterhub API version is not new enough the behavior is
the same and we filter client-side.

Details on manual testing can be found in the issue comments.

[1] https://jupyterhub.readthedocs.io/en/stable/_static/rest-api/index.html#operation--users-get

Closes #21